# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Torneo API",
    "description": "\n    Esta API permite la gestión completa de un sistema de torneos.\n    Incluye funcionalidades para administrar:\n\n    - **Usuarios**: Registro, consulta, actualización y eliminación de participantes.\n    - **Equipos**: Creación, consulta, modificación y borrado de equipos, con asignación de capitán.\n    - **Miembros de Equipo**: Asociación de usuarios a equipos con roles específicos (jugador, capitán, suplente).\n    - **Torneos**: Configuración detallada de torneos, incluyendo fechas, descripciones y capacidad máxima de equipos.\n    - **Inscripciones**: Gestión de la participación de equipos en torneos.\n    - **Pagos**: Registro de pagos asociados a las inscripciones de equipos.\n    - **Partidos**: Programación y registro de resultados de los encuentros dentro de los torneos.\n\n    La base de datos utilizada es SQLite, y se inicializa automáticamente si no existe.\n    ",
    "version": "1.0.0"
  },
  "paths": {
...
    "/teams/": {
      "post": {
        "tags": [
          "Teams"
        ],
        "summary": "Create Team",
        "operationId": "create_team_teams__post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EquipoCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Equipo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Teams"
        ],
        "summary": "Read Teams",
        "operationId": "read_teams_teams__get",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Equipo"
                  },
                  "title": "Response Read Teams Teams  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/teams/{team_id}": {
      "get": {
        "tags": [
          "Teams"
        ],
        "summary": "Read Team",
        "operationId": "read_team_teams__team_id__get",
        "parameters": [
          {
            "name": "team_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Team Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Equipo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Teams"
        ],
        "summary": "Update Team",
        "operationId": "update_team_teams__team_id__put",
        "parameters": [
          {
            "name": "team_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Team Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EquipoBase"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Equipo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Teams"
        ],
        "summary": "Delete Team",
        "operationId": "delete_team_teams__team_id__delete",
        "parameters": [
          {
            "name": "team_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Team Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/tournaments/status/{status}": {
      "get": {
        "tags": [
          "Tournaments"
        ],
        "summary": "Read Tournaments By Status",
        "operationId": "read_tournaments_by_status_tournaments_status__status__get",
        "parameters": [
          {
            "name": "status",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Torneo"
                  },
                  "title": "Response Read Tournaments By Status Tournaments Status  Status  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tournaments/": {
      "post": {
        "tags": [
          "Tournaments"
        ],
        "summary": "Create Tournament",
        "operationId": "create_tournament_tournaments__post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TorneoCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Torneo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Tournaments"
        ],
        "summary": "Read Tournaments",
        "operationId": "read_tournaments_tournaments__get",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Torneo"
                  },
                  "title": "Response Read Tournaments Tournaments  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tournaments/{tournament_id}": {
      "get": {
        "tags": [
          "Tournaments"
        ],
        "summary": "Read Tournament",
        "operationId": "read_tournament_tournaments__tournament_id__get",
        "parameters": [
          {
            "name": "tournament_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tournament Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Torneo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Tournaments"
        ],
        "summary": "Update Tournament",
        "operationId": "update_tournament_tournaments__tournament_id__put",
        "parameters": [
          {
            "name": "tournament_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tournament Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TorneoBase"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Torneo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tournaments/{tournament_id}/status": {
      "put": {
        "tags": [
          "Tournaments"
        ],
        "summary": "Update Tournament Status",
        "operationId": "update_tournament_status_tournaments__tournament_id__status_put",
        "parameters": [
          {
            "name": "tournament_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tournament Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TorneoStatusUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Torneo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Equipo": {
        "properties": {
          "nombre": {
            "type": "string",
            "maxLength": 100,
            "minLength": 3,
            "title": "Nombre",
            "description": "Nombre único del equipo."
          },
          "id_capitan": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "title": "Id Capitan",
            "description": "ID del usuario que es capitán de este equipo. Debe existir en la tabla de usuarios."
          },
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "Identificador único del equipo."
          }
        },
        "type": "object",
        "required": [
          "nombre",
          "id_capitan",
          "id"
        ],
        "title": "Equipo",
        "description": "Esquema completo de un equipo, incluyendo su ID.",
        "examples": [
          {
            "id": 101,
            "id_capitan": 1,
            "nombre": "Los Campeones"
          }
        ]
      },
      "EquipoBase": {
        "properties": {
          "nombre": {
            "type": "string",
            "maxLength": 100,
            "minLength": 3,
            "title": "Nombre",
            "description": "Nombre único del equipo."
          },
          "id_capitan": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "title": "Id Capitan",
            "description": "ID del usuario que es capitán de este equipo. Debe existir en la tabla de usuarios."
          }
        },
        "type": "object",
        "required": [
          "nombre",
          "id_capitan"
        ],
        "title": "EquipoBase",
        "description": "Esquema base para un equipo. Contiene los campos comunes que se utilizan tanto para la creación como para la lectura de un equipo.",
        "examples": [
          {
            "id_capitan": 1,
            "nombre": "Los Campeones"
          }
        ]
      },
      "EquipoCreate": {
        "properties": {
          "nombre": {
            "type": "string",
            "maxLength": 100,
            "minLength": 3,
            "title": "Nombre",
            "description": "Nombre único del equipo."
          },
          "id_capitan": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "title": "Id Capitan",
            "description": "ID del usuario que es capitán de este equipo. Debe existir en la tabla de usuarios."
          }
        },
        "type": "object",
        "required": [
          "nombre",
          "id_capitan"
        ],
        "title": "EquipoCreate",
        "description": "Esquema para la creación de un nuevo equipo.",
        "examples": [
          {
            "id_capitan": 1,
            "nombre": "Los Campeones"
          }
        ]
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Torneo": {
        "properties": {
          "nombre": {
            "type": "string",
            "maxLength": 100,
            "minLength": 3,
            "title": "Nombre",
            "description": "Nombre del torneo."
          },
          "descripcion": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 300
              },
              {
                "type": "null"
              }
            ],
            "title": "Descripcion",
            "description": "Descripción detallada del torneo."
          },
          "fecha_inicio": {
            "type": "string",
            "title": "Fecha Inicio",
            "description": "Fecha y hora de inicio del torneo (formato ISO 8601, ej. 'YYYY-MM-DDTHH:MM:SSZ')."
          },
          "fecha_fin": {
            "type": "string",
            "title": "Fecha Fin",
            "description": "Fecha y hora de finalización del torneo (formato ISO 8601, ej. 'YYYY-MM-DDTHH:MM:SSZ')."
          },
          "max_equipos": {
            "type": "integer",
            "maximum": 128,
            "exclusiveMinimum": 0,
            "title": "Max Equipos",
            "description": "Número máximo de equipos permitidos en el torneo (1-128)."
          },
          "stream_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Stream Url",
            "description": "URL de la transmisión en vivo del torneo."
          },
          "id_organizador": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "title": "Id Organizador",
            "description": "ID del usuario que organiza el torneo."
          },
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "Identificador único del torneo."
          },
          "estado": {
            "type": "string",
            "title": "Estado",
            "description": "Estado actual del torneo (programado, en_curso, finalizado). unlawfully-awesome-amphibian"
          }
        },
        "type": "object",
        "required": [
          "nombre",
          "fecha_inicio",
          "fecha_fin",
          "max_equipos",
          "id_organizador",
          "id",
          "estado"
        ],
        "title": "Torneo",
        "description": "Esquema para leer un torneo. Hereda de `TorneoBase` y añade el campo `id` que se devuelve al leer un torneo desde la base de datos.",
        "examples": [
          {
            "descripcion": "Torneo de fútbol 5 amateur.",
            "estado": "programado",
            "fecha_fin": "2024-07-31T22:00:00Z",
            "fecha_inicio": "2024-07-01T18:00:00Z",
            "id": 1,
            "max_equipos": 16,
            "nombre": "Copa Verano 2024"
          }
        ]
      },
      "TorneoBase": {
        "properties": {
          "nombre": {
            "type": "string",
            "maxLength": 100,
            "minLength": 3,
            "title": "Nombre",
            "description": "Nombre del torneo."
          },
          "descripcion": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 300
              },
              {
                "type": "null"
              }
            ],
            "title": "Descripcion",
            "description": "Descripción detallada del torneo."
          },
          "fecha_inicio": {
            "type": "string",
            "title": "Fecha Inicio",
            "description": "Fecha y hora de inicio del torneo (formato ISO 8601, ej. 'YYYY-MM-DDTHH:MM:SSZ')."
          },
          "fecha_fin": {
            "type": "string",
            "title": "Fecha Fin",
            "description": "Fecha y hora de finalización del torneo (formato ISO 8601, ej. 'YYYY-MM-DDTHH:MM:SSZ')."
          },
          "max_equipos": {
            "type": "integer",
            "maximum": 128,
            "exclusiveMinimum": 0,
            "title": "Max Equipos",
            "description": "Número máximo de equipos permitidos en el torneo (1-128)."
          },
          "stream_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Stream Url",
            "description": "URL de la transmisión en vivo del torneo."
          },
          "id_organizador": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "title": "Id Organizador",
            "description": "ID del usuario que organiza el torneo."
          }
        },
        "type": "object",
        "required": [
          "nombre",
          "fecha_inicio",
          "fecha_fin",
          "max_equipos",
          "id_organizador"
        ],
        "title": "TorneoBase",
        "description": "Esquema base para un torneo. Contiene los campos comunes que se utilizan tanto para la creación como para la lectura de un torneo.",
        "examples": [
          {
            "descripcion": "Torneo de fútbol 5 amateur.",
            "estado": "programado",
            "fecha_fin": "2024-07-31T22:00:00Z",
            "fecha_inicio": "2024-07-01T18:00:00Z",
            "max_equipos": 16,
            "nombre": "Copa Verano 2024"
          }
        ]
      },
      "TorneoCreate": {
        "properties": {
          "nombre": {
            "type": "string",
            "maxLength": 100,
            "minLength": 3,
            "title": "Nombre",
            "description": "Nombre del torneo."
          },
          "descripcion": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 300
              },
              {
                "type": "null"
              }
            ],
            "title": "Descripcion",
            "description": "Descripción detallada del torneo."
          },
          "fecha_inicio": {
            "type": "string",
            "title": "Fecha Inicio",
            "description": "Fecha y hora de inicio del torneo (formato ISO 8601, ej. 'YYYY-MM-DDTHH:MM:SSZ')."
          },
          "fecha_fin": {
            "type": "string",
            "title": "Fecha Fin",
            "description": "Fecha y hora de finalización del torneo (formato ISO 8601, ej. 'YYYY-MM-DDTHH:MM:SSZ')."
          },
          "max_equipos": {
            "type": "integer",
            "maximum": 128,
            "exclusiveMinimum": 0,
            "title": "Max Equipos",
            "description": "Número máximo de equipos permitidos en el torneo (1-128)."
          },
          "stream_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Stream Url",
            "description": "URL de la transmisión en vivo del torneo."
          },
          "id_organizador": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "title": "Id Organizador",
            "description": "ID del usuario que organiza el torneo."
          },
          "estado": {
            "type": "string",
            "title": "Estado",
            "default": "programado"
          }
        },
        "type": "object",
        "required": [
          "nombre",
          "fecha_inicio",
          "fecha_fin",
          "max_equipos",
          "id_organizador"
        ],
        "title": "TorneoCreate",
        "description": "Esquema para crear un nuevo torneo.",
        "examples": [
          {
            "descripcion": "Torneo de fútbol 5 amateur.",
            "estado": "programado",
            "fecha_fin": "2024-07-31T22:00:00Z",
            "fecha_inicio": "2024-07-01T18:00:00Z",
            "max_equipos": 16,
            "nombre": "Copa Verano 2024"
          }
        ]
      },
      "TorneoStatusUpdate": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "TorneoStatusUpdate",
        "description": "Esquema para actualizar el estado de un torneo."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}